requirements.txt
flask
flask-cors
openai
pyngrok
pandas
sentence-transformers
faiss-cpu
gunicorn
